Move HTML launch logic from View into HtmlDocument
