+++ packages/dioxus-blitz/src/documents/html_document.rs
@@ -1,9 +1,11 @@
+use blitz::Viewport;
 use blitz_dom::Document;
-use super::DocumentLike;
 
+use crate::Config;
+use crate::DocumentLike;
 
 pub struct HtmlDocument {
-    inner: Document
+    inner: Document,
 }
 
 // Implement DocumentLike and required traits for HtmlDocument
@@ -24,3 +26,26 @@ impl Into<Document> for HtmlDocument {
     }
 }
 impl DocumentLike for HtmlDocument {}
+
+impl HtmlDocument {
+    pub(crate) fn from_html(html: &str, cfg: &Config) -> Self {
+        // Spin up the virtualdom and include the default stylesheet
+        let mut dom = Document::new(Viewport::new((0, 0)).make_device());
+
+        // Set base url if configured
+        if let Some(url) = &cfg.base_url {
+            dom.set_base_url(&url);
+        }
+
+        // Include default and user-specified stylesheets
+        dom.add_stylesheet(include_str!("./default.css"));
+        for ss in &cfg.stylesheets {
+            dom.add_stylesheet(&ss);
+        }
+
+        // Populate dom with HTML
+        dom.write(&html);
+
+        HtmlDocument { inner: dom }
+    }
+}
